[config] move profile creation to georestheart lib
