+++ src/main/scala/se/lu/nateko/cp/cpauth/accounts/RestHeartClient.scala
@@ -44,70 +44,6 @@ class RestHeartClient(
 	private val KeepIdsOnly = "keys" -> "{\"_id\": 1}"
 	private def pageSizeQpar(size: Int) = "pagesize" -> size.toString
 
-	def createUserIfNew(uid: UserId, givenName: String, surname: String)(using Envri): Future[Done] =
-		userExists(uid).flatMap{
-			exists => if(exists) ok else createNewUser(uid, givenName, surname)
-		}.andThen{
-			case Failure(exc) => log.error(exc, s"Problem creating user ${uid.email}")
-		}
-
-	private def createNewUser(uid: UserId, givenName: String, surname: String)(using Envri): Future[Done] =
-		val payload = JsObject(
-			"_id" -> JsString(uid.email),
-			"profile" -> JsObject(
-				"givenName" -> JsString(givenName),
-				"surname" -> JsString(surname)
-			),
-			"cart" -> JsObject(
-				"_items" -> JsArray.empty
-			)
-		)
-		val createStatus = for(
-			entity <- Marshal(payload).to[RequestEntity];
-			req = HttpRequest(
-				method = HttpMethods.POST,
-				uri = config.usersCollUri,
-				entity = entity,
-				headers = Seq(headers.`Content-Type`(ContentTypes.`application/json`))
-			);
-			status <- requestDiscardResp(req)
-		) yield status
-
-		createStatus.flatMap{status =>
-			if(status.isSuccess()) ok
-			else fail(
-				s"Could not create user ${uid.email}, got response ${status.defaultMessage()}"
-			)
-		}
-
-	def userExists(uid: UserId)(using Envri): Future[Boolean] = {
-		val query = Uri.Query(KeepIdsOnly)
-		val req = HttpRequest(uri = getUserUri(uid).withQuery(query))
-
-		requestDiscardResp(req).flatMap{status =>
-			if(status.isSuccess()) Future.successful(true)
-			else if(status == StatusCodes.NotFound) Future.successful(false)
-			else fail(
-				s"Failed checking user ${uid.email}, got response " + status.defaultMessage()
-			)
-		}
-	}
-
-
-	def getGivenAndSurName(uid: UserId)(using Envri): Future[(String, String)] =
-		getUserProps(uid, Seq("profile.givenName", "profile.surname")).flatMap{userObj =>
-
-			val profileVal = getFieldOpt[JsValue](userObj, "profile").getOrElse(JsObject.empty);
-
-			Future.fromTry(
-				for(
-					profile <- ensure[JsObject](profileVal);
-					givenName <- getStringField(profile, "givenName");
-					surname <- getStringField(profile, "surname")
-				) yield (givenName, surname)
-			)
-		}
-
 	def findUsers(filter: Map[String, String])(using Envri): Future[Seq[UserId]] = {
 
 		val filterParam: String = filter.map{
